requirements.txt

Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
cryptography==41.0.7
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.26.5
django-filter==23.3
redis==5.0.1
celery==5.3.4